refactor: extract list transactions use case
